refactor(desktop): group worker threads into main/workers
