+++ apps/desktop/src/main/ipc/waveform.ts
@@ -6,7 +6,7 @@ import { IPC_CHANNELS, WAVEFORM_PEAK_COUNT } from '@shiranami/contracts';
 import { handle } from './with-ipc-handler';
 import { logger } from '../app/logger';
 import { hashTrackKey, readCachedPeaks, writeCachedPeaks } from '../shared/waveform-cache';
-import { decodeWaveformPeaks, shutdownWaveformWorker } from '../waveform-host';
+import { decodeWaveformPeaks, shutdownWaveformWorker } from '../workers/waveform-host';
 import { waveformGetPeaksArgs } from './schemas/waveform';
 
 const C = IPC_CHANNELS.waveform;
