+++ apps/desktop/src/main/workers/waveform-worker.ts
